tasks/make_boundaries/pyproject.toml
[project]
name = "fetch-geometry"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dotenv>=0.9.9",
    "geopandas>=1.1.2",
    "google-cloud-storage>=3.9.0",
    "pandas>=3.0.1",
    "requests>=2.32.5",
    "shapely>=2.1.2",
]
